+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 -   Maximum width on select widgets.
 -   Improved position of reference ID in edit mode.
 -   Fixed editability of code blocks.
+-   Allow chapter navigation to forthcoming chapters if editable.
 
 # 0.5.63 - 2023-11-11